# microservice/src/main/resources/swagger/api.yml
# API-first development with OpenAPI
# This file will be used at compile time to generate Spring-MVC endpoint stubs using openapi-generator
openapi: '3.0.1'
info:
  title: 'microservice'
  version: 0.0.1
servers:
  - url: http://localhost:8081/api
    description: Development server
  - url: https://localhost:8081/api
    description: Development server with TLS Profile
paths:
  /client:
    post:
      summary: Add a new client
      description: Registers a new Client in the system
      operationId: putClient
      requestBody:
        description: The Client to add
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientDTO'
      responses:
        '200':
          description: Client added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientDTO'
components:
  schemas:
    ClientDTO:
      type: object
      properties:
        name:
          type: string
        nif:
          type: string
        address:
          $ref: '#/definitions/AddressDTO'
        phone:
          $ref: '#/definitions/PhoneNumberDTO'
      required:
        - name
        - nif
        - address
        - phone
    AddressDTO:
      description: The textual identifier of the location where the client resides at
      type: object
      properties:
        city:
          description: City
          type: string
        country:
          description: Country
          type: number
        postcode:
          description: Postcode
          type: string
        stateOrProvince:
          description: State or province
          type: string
        streetOne:
          description: Describes the street
          type: string
        streetTwo:
          description: Complementary street description
          type: string
        emailAddress:
          description: Full email address in standard format
          type: string
      required:
        - city
        - country
        - postcode
        - stateOrProvince
        - streetOne
        - streetTwo
        - emailAddress
    PhoneNumberDTO:
      description: The phone number with country code attributed to this client and this client alone
      type: object
      properties:
        countryCode:
          description: The code which uniquely identifies the country of this phone number
          type: number
        number:
          description: The number that, within the country with the provided countryCode, belongs solely to the client
          type: number
      required:
        - countryCode
        - number
  responses:
    Problem:
      description: error occurred - see status code and problem object for more information.
      content:
        application/problem+json:
          schema:
            type: object
            properties:
              type:
                type: string
                format: uri
                description: |
                  An absolute URI that identifies the problem type.  When dereferenced,
                  it SHOULD provide human-readable documentation for the problem type
                  (e.g., using HTML).
                default: 'about:blank'
                example: 'https://www.jhipster.tech/problem/problem-with-message'
              title:
                type: string
                description: |
                  A short, summary of the problem type. Written in english and readable
                  for engineers (usually not suited for non technical stakeholders and
                  not localized); example: Service Unavailable
              status:
                type: integer
                format: int32
                description: |
                  The HTTP status code generated by the origin server for this occurrence
                  of the problem.
                minimum: 100
                maximum: 600
                exclusiveMaximum: true
                example: 503
              detail:
                type: string
                description: |
                  A human readable explanation specific to this occurrence of the
                  problem.
                example: Connection to database timed out
              instance:
                type: string
                format: uri
                description: |
                  An absolute URI that identifies the specific occurrence of the problem.
                  It may or may not yield further information if dereferenced.

  securitySchemes:
    jwt:
      type: http
      description: JWT Authentication
      scheme: bearer
      bearerFormat: JWT
#    basic:
#      type: http
#      description: Basic Authentication
#      scheme: basic
security:
  - jwt: []
#  - basic: []